edit type of timeWAit in CompleteForms.py and GetElement.py

+++ CompleteForms.py
@@ -10,23 +10,23 @@ class CompleteForms(GetElement):
     def __init__(self, URL):
         GetElement.__init__(self, URL)
 
-    def completeSelectForm(self, value, element, el_type=By.ID, timeWait: int = 60):
+    def completeSelectForm(self, value, element, el_type=By.ID, timeWait: float = 60):
         selectFrom = Select(self.wait_and_return(element, el_type, timeWait))
         selectFrom.select_by_value(value)
 
-    def completeInputForm(self, value, element, el_type=By.ID, timeWait: int = 60):
+    def completeInputForm(self, value, element, el_type=By.ID, timeWait: float = 60):
         inputForm = self.wait_and_return(element, el_type, timeWait)
         inputForm.clear()
         inputForm.send_keys(value)
 
-    def complete_input_scroll(self, value, element, el_type=By.ID, timeSleep: float = 0.5, timeWait: int = 60):
+    def complete_input_scroll(self, value, element, el_type=By.ID, timeSleep: float = 0.5, timeWait: float = 60):
         inputForm = self.wait_and_return(element, el_type, timeWait)
         inputForm.location_once_scrolled_into_view
         time.sleep(timeSleep)
         inputForm.clear()
         inputForm.send_keys(value)
 
-    def completeInputFormReadOnly(self, value, element, el_type=By.ID, timeWait: int = 60):
+    def completeInputFormReadOnly(self, value, element, el_type=By.ID, timeWait: float = 60):
         inputForm = self.wait_and_return(element, el_type, timeWait)
         selector: str = ""
         isMulty: bool = False
@@ -53,23 +53,23 @@ class CompleteForms(GetElement):
         inputForm.clear()
         inputForm.send_keys(value)
 
-    def completeRadio(self, value, element, el_type=By.ID, timeWait: int = 60):
+    def completeRadio(self, value, element, el_type=By.ID, timeWait: float = 60):
         radioForm = self.wait_and_returns(element, el_type, timeWait)
         radioForm[value].click()
 
-    def completeCheckBox(self, element, el_type=By.ID, timeWait: int = 60):
+    def completeCheckBox(self, element, el_type=By.ID, timeWait: float = 60):
         checkBoxForm = self.wait_and_return(element, el_type, timeWait)
         checkBoxForm.click()
 
-    def clickBtn(self, element, el_type=By.ID, timeWait: int = 60):
+    def clickBtn(self, element, el_type=By.ID, timeWait: float = 60):
         pageLink = self.wait_and_return(element, el_type, timeWait)
         pageLink.click()
 
-    def click_btn_on_list(self, element, index: int, el_type=By.ID, timeWait: int = 60):
+    def click_btn_on_list(self, element, index: int, el_type=By.ID, timeWait: float = 60):
         btns = self.wait_and_returns(element, el_type, timeWait)
         btns[index].click()
 
-    def click_btn_scroll(self, element, el_type=By.ID, timeSleep: float = 0.5, timeWait: int = 60):
+    def click_btn_scroll(self, element, el_type=By.ID, timeSleep: float = 0.5, timeWait: float = 60):
         btn = self.wait_and_return(element, el_type, timeWait)
         btn.location_once_scrolled_into_view
         time.sleep(timeSleep)
